requirements.txt
uvicorn==0.13.4
requests==2.25.1
requests_toolbelt==0.9.1
fastapi==0.63.0
python-binance==0.7.10
TA-Lib
python-multipart
numpy
sqlalchemy
psycopg2-binary